+++ src/main/java/nextstep/auth/ui/AuthenticationPrincipalArgumentResolver.java
@@ -1,7 +1,7 @@
-package nextstep.member.ui;
+package nextstep.auth.ui;
 
-import nextstep.member.AuthenticationException;
-import nextstep.member.application.JwtTokenProvider;
+import nextstep.auth.AuthenticationException;
+import nextstep.auth.application.JwtTokenProvider;
 import nextstep.member.domain.LoginMember;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
