+++ src/main/java/nextstep/auth/ui/TokenController.java
@@ -1,9 +1,9 @@
-package nextstep.member.ui;
+package nextstep.auth.ui;
 
 import nextstep.favorite.application.dto.GithubAccessTokenRequest;
-import nextstep.member.application.TokenService;
-import nextstep.member.application.dto.TokenRequest;
-import nextstep.member.application.dto.TokenResponse;
+import nextstep.auth.application.TokenService;
+import nextstep.auth.application.dto.TokenRequest;
+import nextstep.auth.application.dto.TokenResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
