+++ src/main/java/nextstep/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package nextstep.exception;
 
-import nextstep.member.AuthenticationException;
+import nextstep.auth.AuthenticationException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
